extract mesh_cq interface to hide its details of impl
