+++ tt_metal/distributed/CMakeLists.txt
@@ -2,7 +2,7 @@ set(DISTRIBUTED_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/coordinate_translation.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/distributed.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mesh_buffer.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_command_queue.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/hardware_mesh_command_queue.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mesh_device.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mesh_device_view.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mesh_event.cpp
